[Fix] #143 - 회원 탈퇴 변수 자료형 수정

+++ src/main/java/shop/cazait/domain/master/api/MasterController.java
@@ -98,7 +98,7 @@ public class MasterController {
 
 	@DeleteMapping
 	@ApiOperation(value = "마스터 계정 탈퇴(상태  변경)", notes = "특정 ID의 마스터 상태를 INACTIVE로 변경한다.")
-	public SuccessResponse<String> deleteMaster(@Validated @PathVariable int id) throws MasterException {
+	public SuccessResponse<String> deleteMaster(@Validated @PathVariable Long id) throws MasterException {
 		masterService.removeMaster(id);
 		String response = "회원 탈퇴가 성공하였습니다.";
 		return new SuccessResponse<>(response);
